% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnS3.R
\name{knn_s3}
\alias{knn_s3}
\title{k-NN Classification and Regression}
\usage{
knn_s3(formula, data, k = 5L)
}
\arguments{
\item{formula}{n object of class \code{\link{formula}}, y ~ x1 + x2 + ...}

\item{data}{a data frame containing the variables used in the formula.}

\item{k}{number of neighbors}
}
\value{
An object of \code{\link{class}} \code{knn_s3} containing the
following components:
the original function call, the used formula, values of features, the values of
the response variable, the model type (classification or regression),
and the number of neighbors \code{k}.
}
\description{
Create a k-nearest neighbors (k-NN) model object for classification or regression
using a formula interface. This function supports numeric responses for regression
and factor responses for classification.
}
\details{
The response variable should be either numeric (for regression) or a factor (for classification).
The function throws an error if the response is not one of these types.
}
\examples{
\dontrun{
data(mtcars)
fit <- knn_s3(mpg ~ disp + hp, data = mtcars, k = 3)
class(fit)
}

}
