% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.knn_s3}
\alias{plot.knn_s3}
\title{Plot method for knn_s3 objects}
\usage{
\method{plot}{knn_s3}(object, ...)
}
\arguments{
\item{object}{A fitted \code{\link{knn_s3}} object.}

\item{...}{further arguments.}
}
\value{
Returns an object of \code{\link{ggplot}} class visualizing model predictions
 against actual values.
}
\description{
Produces a plot comparing actual versus predicted values for a \code{knn_s3} model.
For regression models, it shows a scatter plot and For classification models, it
creates a bar plot of counts for each actual class filled by predicted classes.
}
\examples{
\dontrun{
fit <- knn_s3(Species ~ Sepal.Length + Sepal.Width, data = iris, k = 5)
plot(fit)
}
}
