% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdist.R
\name{plotdist.knn_s3}
\alias{plotdist.knn_s3}
\title{Plot histogram of distances in knn_s3 training data}
\usage{
plotdist.knn_s3(object, distmethod = c("sse", "sad"))
}
\arguments{
\item{object}{A fitted \code{\link{knn_s3}} object.}

\item{distmethod}{A character string specifying the distance method to use.
One of \code{"sse"} (sum of squared errors) or \code{"sad"} (sum of absolute differences).}
}
\value{
Returns an object of \code{\link{ggplot}} class showing the histogram of distances.
}
\description{
Plot histogram of distances in knn_s3 training data
}
\examples{
\dontrun{
fit <- knn_s3(mpg ~ disp + hp, data = mtcars, k = 3)
plotdist.knn_s3(fit, distmethod = "sse")
}
}
