% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.knn_s3}
\alias{print.knn_s3}
\title{Print method for knn_s3 objects}
\usage{
\method{print}{knn_s3}(object, ...)
}
\arguments{
\item{object}{an object of class (\code{\link{knn_s3}}).}

\item{...}{further arguments.}
}
\value{
Prints the \code{knn_s3} object structure and key attributes.
}
\description{
Print the model structure and key attributes of a \code{\link{knn_s3}} object.
}
\examples{
\dontrun{
# knn regression model
data(mtcars)
fit_reg <- knn_s3(mpg ~ disp + hp, data = mtcars, k = 3)
print(fit_reg)

# knn classification model
data(iris)
fit_class <- knn_s3(Species ~ Sepal.Length + Sepal.Width, data = iris, k = 3)
print(fit_class)
}
}
