% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.knn_s3}
\alias{summary.knn_s3}
\title{Summarizing knn Classification and Regression Model Fits}
\usage{
\method{summary}{knn_s3}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{knn_s3}}.}
}
\value{
A summary of error metrics for \code{knn_s3} Model.
\itemize{
 \item For regression: \code{MSE} (mean squared error) and \code{R2} (coefficient of determination).
 \item For classification: \code{Accuracy} and \code{ConfusionMatrix} table.
}
}
\description{
Provides a summary of a fitted knn Classification and Regression Model,
including error metrics for regression or accuracy and confusion matrix
for classification.
}
\examples{
\dontrun{
# knn regression model
data(mtcars)
fit_reg <- knn_s3(mpg ~ disp + hp, data = mtcars, k = 3)
print(fit_reg)
summary(fit_reg)

# knn classification model
data(iris)
fit_class <- knn_s3(Species ~ Sepal.Length + Sepal.Width, data = iris, k = 3)
print(fit_class)
summary(fit_class)
}
}
